+++ charbot/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+from .bot import CBot
+
+
+__all__ = ("CBot",)